+++ SolastaCommunityExpansion/Models/CustomReactionsContext.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using SolastaCommunityExpansion.Api.AdditionalExtensions;
-using SolastaCommunityExpansion.CustomDefinitions;
+using SolastaCommunityExpansion.CustomUI;
 using SolastaModApi.Infrastructure;
 using static ActionDefinitions;
 using static ActionDefinitions.ActionStatus;
